+++ src/core/router/engines/browser-history.ts
@@ -140,6 +140,9 @@ export default function createRouter(component: bRouter): Router {
 		newListener: false
 	});
 
+	let currentLocation;
+	saveLocation();
+
 	const router = Object.mixin({withDescriptors: 'onlyAccessors'}, Object.create(emitter), {
 		get route(): CanUndef<Route> {
 			const
@@ -304,8 +307,15 @@ export default function createRouter(component: bRouter): Router {
 	});
 
 	$a.on(globalThis, 'popstate', async () => {
+		const prevLocation = currentLocation;
+		saveLocation();
+
 		if (browser.is.iOS !== false && isOpenedFromBFCache) {
 			isOpenedFromBFCache = false;
+			return;
+		}
+
+		if (onlyHashChange(prevLocation, currentLocation)) {
 			return;
 		}
 
@@ -340,7 +350,26 @@ export default function createRouter(component: bRouter): Router {
 
 	return router;
 
+	function saveLocation() {
+		currentLocation = new URL(location.href);
+	}
+
+	function onlyHashChange(location1: URL, location2: URL) {
+		if (location1.hash !== '' || location2.hash !== '') {
+			const
+				hashRgxp = /#.*/;
+
+			if (location1.href.replace(hashRgxp, '') === location2.href.replace(hashRgxp, '')) {
+				return true;
+			}
+		}
+
+		return false;
+	}
+
 	function load(route: string, params?: Route, method: string = 'pushState'): Promise<void> {
+		params = Object.fastClone(params, {freezable: false});
+
 		if (!Object.isTruly(route)) {
 			throw new ReferenceError('The page to load is not specified');
 		}
@@ -411,6 +440,7 @@ export default function createRouter(component: bRouter): Router {
 				// to avoid DataCloneError we have to clone it with `Object.mixin({deep: true})`
 				const filteredParams = Object.mixin({deep: true, filter: (el) => !Object.isFunction(el)}, {}, params);
 				history[method](filteredParams, params.name, route);
+				saveLocation();
 			}
 
 			const